cleanup even more stuff from the game module that just wrapped/re-exported
