--- chers/src/moves/mod.rs
@@ -1,2 +1,0 @@
-pub mod serialization;
-pub mod transport;
