+++ chers_cli/src/moves/transport.rs
@@ -2,36 +2,28 @@ use std::error::Error;
 use std::io::Read;
 use std::{io::Write, net::TcpStream};
 
-use crate::Move;
+use chers::Move;
 
 use super::serialization::Converter;
 
-/// Converts moves from and into different string representations.
 pub trait Transport {
-    /// Send [`a_move`] to the other party
     fn send(&mut self, a_move: &Move) -> Result<(), Box<dyn Error>>;
-
-    /// Wait and block, until the other party has made their move.
     fn receive(&mut self) -> Result<Move, Box<dyn Error>>;
 }
 
-/// Juggles transferring moves and updating game state.
 pub struct TcpTransport {
     stream: TcpStream,
     converter: Box<dyn Converter>,
 }
 
 impl Transport for TcpTransport {
-    /// Sends a local move to the remote party.
     fn send(&mut self, a_move: &Move) -> Result<(), Box<dyn Error>> {
         let serialized = self.converter.serialize(a_move);
 
         Ok(self.stream.write_all(serialized.as_bytes())?)
     }
 
-    /// Waits for the remote party to make their move.
     fn receive(&mut self) -> Result<Move, Box<dyn Error>> {
-        // 128 should be more than enough to serialize a simple move
         let mut buffer: [u8; 128] = [0; 128];
         match self.stream.read_exact(&mut buffer) {
             Ok(it) => it,
@@ -44,7 +36,6 @@ impl Transport for TcpTransport {
     }
 }
 
-/// Juggles transferring moves and updating game state.
 pub struct Coordinator {
     stream: TcpStream,
     converter: Box<dyn Converter>,
@@ -55,16 +46,13 @@ impl Coordinator {
         Self { stream, converter }
     }
 
-    /// Sends a local move to the remote party.
     pub fn send(&mut self, a_move: &Move) -> Result<(), std::io::Error> {
         let serialized = self.converter.serialize(a_move);
 
         self.stream.write_all(serialized.as_bytes())
     }
 
-    /// Waits for the remote party to make their move.
     pub fn receive(&mut self) -> Result<Move, Box<dyn Error>> {
-        // 128 should be more than enough to serialize a simple move
         let mut buffer: [u8; 128] = [0; 128];
         self.stream.read_exact(&mut buffer)?;
